chore(compose): factor renderSinglePattern into its own module
wip
